2단계 - 지하철 노선 관리
- distance 컬럼 추가
- 불필요 컴포넌트 제거
- @Column 추가

+++ src/main/java/subway/line/LineRequest.java
@@ -22,7 +22,7 @@ public class LineRequest {
 
     private Long downStationId;
 
-    private Long distance;
+    private Integer distance;
 
     @JsonIgnore
     public List<Long> getStationIds() {
